Adjusted Alignments & Code Cleanup
Adjusted where the L & R lined up on each key, created separate ::Before & ::After for the note-labels based on key color.   Aligned them so that it was easy to read and had enough space for the fingerings once I get that implemented.

Organized, commented, and adjusted the Keyboards.JS file so that it is easier to track what is happening where.