Refactoring code to be less manhandled
Realized that some of my matching functions aren't returning sharps properly during a refactor.

Most of the issue seems to be with converting from the symbol code for sharp and matching to a #.  Need to sort a conversion properly.

Have it happen elsewhere in the code will need to review where that is.  Don't have time at the moment.