Adjusted program flow
Pulled the deletePianoKeyboard and generate Keyboard and vexFlow out of the setSelectedScaleButton function into an updateKeyboard function.

Refactored the button click for key and scale so that the setSelected scale and key functions did not directly call keyboard update.  This was causing multiple unrequired calls and eventually a stack overflow due to infinite loop.